pyproject.toml
[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "arviz-base"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
    {name = "ArviZ team", email = "arvizdevs@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dynamic = ["version", "description"]
dependencies = [
  "numpy>=1.23",
  "xarray>=2022.6.0",
  "xarray-datatree",
  "typing-extensions>=3.10",
]

[tool.flit.module]
name = "arviz_base"

[project.urls]
source = "https://github.com/arviz-devs/arviz-base"
tracker = "https://github.com/arviz-devs/arviz-base/issues"
documentation = "https://arviz-base.readthedocs.io"
funding = "https://opencollective.com/arviz"

[project.optional-dependencies]
test = [
    "hypothesis",
    "pytest",
    "pytest-cov",
]
ci = [
    "pre-commit",
    "cloudpickle"
]
doc = [
    "sphinx-book-theme",
    "myst-parser[linkify]",
    "myst-nb",
    "sphinx-copybutton",
    "numpydoc",
    "sphinx>=5",
    "sphinx-design",
    "jupyter-sphinx",
    "netcdf4"
]
netcdf4 = [
  "netcdf4"
]
h5netcdf = [
  "h5netcdf"
]
zarr = [
  "zarr"
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "F",  # Pyflakes
    "E",  # Pycodestyle
    "W",  # Pycodestyle
    "D",  # pydocstyle
    "NPY", # numpy specific rules
    "UP",  # pyupgrade
    "I",  # isort
    "PL",  # Pylint
]
ignore = [
    "PLR0912",  # too many branches
    "PLR0913",  # too many arguments
    "PLR2004",  # magic value comparison
]

[tool.ruff.lint.per-file-ignores]
"src/arviz_base/__init__.py" = ["I", "F401", "E402", "F403"]
"src/arviz_base/example_data/**/*" = ["F", "E", "W", "D", "I", "PL"]
"tests/**/*" = ["D", "PLR2004"]
"external_tests/**/*" = ["D", "PLR2004"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.format]
docstring-code-format = false

[tool.mypy]
python_version = "3.10"
warn_redundant_casts = true
warn_unused_configs = true
pretty = true
show_error_codes = true
show_error_context = true
show_column_numbers = true

disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true

# More strict checks for library code
[[tool.mypy.overrides]]
module = "arviz_base"
disallow_untyped_defs = true

# Ignore certain missing imports
# [[tool.mypy.overrides]]
# module = "thirdparty.*"
# ignore_missing_imports = true

[tool.pytest.ini_options]
filterwarnings = ["error"]
testpaths = [
    "tests",
]

[tool.coverage.run]
source = ["arviz_base"]
